refactor: amend package structure
