+++ src/main/java/kr/bb/notification/domain/notification/application/NotificationCommandService.java
@@ -5,9 +5,9 @@ import bloomingblooms.domain.notification.PublishNotificationInformation;
 import bloomingblooms.domain.resale.ResaleNotificationList;
 import java.util.List;
 import kr.bb.notification.domain.notification.entity.MemberNotification;
-import kr.bb.notification.domain.notification.entity.MemberNotificationCommand;
+import kr.bb.notification.domain.notification.mapper.MemberNotificationCommand;
 import kr.bb.notification.domain.notification.entity.Notification;
-import kr.bb.notification.domain.notification.entity.NotificationCommand;
+import kr.bb.notification.domain.notification.mapper.NotificationCommand;
 import kr.bb.notification.domain.notification.repository.NotificationJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
