+++ src/main/java/kr/bb/notification/domain/notification/helper/NotificationActionHelper.java
@@ -13,7 +13,7 @@ import bloomingblooms.domain.resale.ResaleNotificationList;
 import java.util.List;
 import kr.bb.notification.common.annotation.OutOfStockNotificationDuplicateCheck;
 import kr.bb.notification.domain.notification.application.NotificationCommandService;
-import kr.bb.notification.domain.notification.entity.NotificationCommand.NotificationInformation;
+import kr.bb.notification.domain.notification.mapper.NotificationCommand.NotificationInformation;
 import kr.bb.notification.domain.notification.infrastructure.sms.SendSMS;
 import kr.bb.notification.domain.notification.infrastructure.sse.SendSSE;
 import lombok.RequiredArgsConstructor;
