+++ src/main/java/kr/bb/notification/domain/notification/infrastructure/action/InfrastructureActionHandler.java
@@ -1,7 +1,7 @@
 package kr.bb.notification.domain.notification.infrastructure.action;
 
 
-import kr.bb.notification.domain.notification.entity.NotificationCommand;
+import kr.bb.notification.domain.notification.mapper.NotificationCommand;
 
 public interface InfrastructureActionHandler<
     T extends NotificationCommand.NotificationInformation> {
