+++ src/main/java/kr/bb/notification/domain/notification/infrastructure/sms/SendSMS.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 import java.util.Map;
 import kr.bb.notification.common.annotation.DuplicateEventCheck;
 import kr.bb.notification.config.AWSConfiguration;
-import kr.bb.notification.domain.notification.entity.NotificationCommand.NotificationInformation;
+import kr.bb.notification.domain.notification.mapper.NotificationCommand.NotificationInformation;
 import kr.bb.notification.domain.notification.infrastructure.action.InfrastructureActionHandler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
