+++ src/main/java/kr/bb/notification/domain/notification/infrastructure/sse/SendSSE.java
@@ -2,7 +2,7 @@ package kr.bb.notification.domain.notification.infrastructure.sse;
 
 import kr.bb.notification.common.annotation.DuplicateEventCheck;
 import kr.bb.notification.domain.emitter.application.SseService;
-import kr.bb.notification.domain.notification.entity.NotificationCommand.NotificationInformation;
+import kr.bb.notification.domain.notification.mapper.NotificationCommand.NotificationInformation;
 import kr.bb.notification.domain.notification.infrastructure.action.InfrastructureActionHandler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
