+++ src/main/java/kr/bb/notification/domain/notification/mapper/NotificationCommand.java
@@ -1,4 +1,4 @@
-package kr.bb.notification.domain.notification.entity;
+package kr.bb.notification.domain.notification.mapper;
 
 import bloomingblooms.domain.notification.NotificationData;
 import bloomingblooms.domain.notification.NotificationKind;
@@ -10,6 +10,8 @@ import bloomingblooms.domain.order.OrderStatusNotification;
 import bloomingblooms.domain.resale.ResaleNotificationList;
 import java.util.List;
 import java.util.stream.Collectors;
+import kr.bb.notification.domain.notification.entity.MemberNotification;
+import kr.bb.notification.domain.notification.entity.Notification;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
