+++ src/test/java/kr/bb/notification/domain/notification/facade/NotificationQueryFacadeHandlerTest.java
@@ -8,8 +8,8 @@ import java.util.ArrayList;
 import java.util.List;
 import kr.bb.notification.domain.notification.entity.MemberNotification;
 import kr.bb.notification.domain.notification.entity.Notification;
-import kr.bb.notification.domain.notification.entity.NotificationCommand.NotificationList;
-import kr.bb.notification.domain.notification.entity.NotificationCommand.UnreadNotificationCount;
+import kr.bb.notification.domain.notification.mapper.NotificationCommand.NotificationList;
+import kr.bb.notification.domain.notification.mapper.NotificationCommand.UnreadNotificationCount;
 import kr.bb.notification.domain.notification.helper.NotificationQueryActionHelper;
 import kr.bb.notification.domain.notification.repository.NotificationJpaRepository;
 import org.junit.jupiter.api.DisplayName;
